feat(ui): add slim HeaderBar (Phase 2)
NIAC's first proper HeaderBar, following the canonical three-slot
convention documented in stem/ui/SHELL.md.

  ┌──────────────────────────────────────────────────────────────┐
  │ [logo] NIAC  [ConnectionStatus]    …    [theme toggle]       │
  └──────────────────────────────────────────────────────────────┘

Per-product slot fills
- LEFT: Network-icon logo on the brand gradient + "NIAC" wordmark
  + ConnectionStatus (existing component, polls /api/v1/version).
- CENTER: empty (reserved for breadcrumbs / page chrome).
- RIGHT: theme toggle only. NIAC has no profiles or interfaces; help
  and settings live in the sidebar footer; there's no logout
  (NIAC is single-user).

App.tsx
- Replaces the inline TopBar function with the new components/HeaderBar.
- Imports + uses `<HeaderBar />` in the SidebarLayout topBar slot.
- Drops Moon/Sun/ConnectionStatus imports (now encapsulated in HeaderBar).
- Drops useTheme + useTranslation('common') imports (no longer needed
  in App.tsx itself; HeaderBar owns them).

Banner identifies this file as per-product (not synced), with a
pointer to the SHELL.md shape convention.
